Function that finds number of bits you'd need to flip to get from one number to the next

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -10,7 +10,7 @@
 
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	int result, i;
+	unsigned long int result, i;
 
 	result = 0;
 	i = n ^ m;
